rename MyHomePage to BingoPage and move it to separate file
